Rename file header-nav.vue to HeaderNav.vue
